GON-25 Add Single Unit Network Interface
The single_unit_network_interface (SUNI) module
enables the connection of a single processing element
to the NoC.

Currently, the SUNI allows the connection of processing
elements that support AXI-Stream interface.

It can be instantiated to support both the AXI-Stream Manager
and AXI-Stream Subordinate interfaces, as well as to support
anyone of them in a standlone mode.