GON-14 Replace Network Data downsizer
This commit replaces temporaly the network data downsizer module content.

It cannot be guaranteed that original code is AXI handshake compatible at
this point. The initial version requires more work to do to satisfy
compatibility with AXI handsheke, but it could be a more optimized
version than the current one.