GON-15 Replace AXIS data upsizer
This commit replaces temporaly the content of the AXIS
data upsizer module.

It has been detected that a deadlock could happen
when the previous module (network_data_downsizer) does not
deliver four consecutive bytes to compose a 32-bit word.

And the previous module must deassert valid after delivering
the third byte of the word in some scenarios.

The current code could not be as effecient as the original,
so more work must be done in the original to solve the deadlock.