GON-46 Add verification environment for NoC
Adds a bunch of SystemVerilog files,
the Xilinx AXI4Stream M_AXIS/S_AXIS VIP
and vivado TCL script to generate a Vivado
project with the Verification environment.

Currently, it is configured to generate
a 4x2 mesh topology where tile 2 has the M_AXIS VIP
and the rest of tiles the S_AXIS VIP. The LBDR routing
method is configured and programmed with XY routing algorithm.

The generator generates AXI4Stream transactions
with TDEST set to zero, so the destination of the
transfers generated by tile 2 is tile 0.