GON-25 Add Reg File Access support to SUNI
This commit adds support to allow that a NoC controller
can access to the register file of the router to write
and read control and status registers, through a
control and status virtual network (default to VN0).

In addition, it provides the updates to the simulation environment to
make it compatible with the SUNI addons. Anyway, the simulation
environment disables the reg file interface, so every
virtual network is treated as a data network. To test register
file access use the NoC Reg File Access Verification environment
provided by GON-56.

SUNI stands for Single Unit Network Interface, which means that
a single unit (processing element) can connect to the NoC through
the NI.