GON-10 Fix lost of flits in net. ejector
The NoC is based on Avail/Valid handshake. In this protocol
Valid can only be asserted when avail is asserted.

When avail is deaserted the NoC component requires two extra
cycles to deassert valid, so a two-slot buffer is required
to avoid lost of flits at reception.

Previously, the network ejector was using two buffers to
overcome the problem, but with some seeds it has been detected
that it is not enough and a three-slot buffer is required.

It is due to the fact that when the output port of the ejector
is busy (and cannot deliver the flit to the next module), then
it reacts deasserting avail signal, but then the router requires
two extra cycles to deassert valid.

On solution would be to deassert avail using combinatorial logic
when valid is asserted, but it could end up on combinatorial cycles
and long combinatorial paths.

The other would be to use three-slot buffer to reduce complexity
to the synthesis tool.

This commit solves the issue using the second option.