GON-46 Adapt NoC simulation environment
This commit adapts the simulation environment of the NoC to the
new additions to the NI for supporting Configuration Register File
Access.

Anyway, the simulation environment disables the FileReg access
interfaces in the NI, so every Virtual Network is used for data.

Besides, several simulation seeds have been tried out for the testbench
and it has been detected that a few of them (seed 13) FAIL in passing the test.

The simulation environment builds a 4x2 mesh where Tile 2 generates
AXI-Stream transactions addressed to Tile 0 using three different
virtual networks.

In <repo>/vivado/2022.2/network_on_chip/scripts there is a TCL script
that can be used for generating the simulation environment.

We plan to improve the simulation environment to make it more robust and
configurable in order to test different randomized configurations
easily.