GON-56 Add Router FileReg Access sim. environment
This commits add the simulation environment to test the Router FileReg
Access.

The environment builds a 4x2 mesh and uses virtual network 0 as a
control and status network.

The control and status network cannot be reconfigured, so the test
prevents to write address 0 of register file, since the design is not
constrained in anycase currently.

TODO forbid to write certain register file addresses.